% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_crit.R
\name{A_crit}
\alias{A_crit}
\title{A_crit}
\usage{
A_crit(M)
}
\arguments{
\item{M}{a matrix object}
}
\value{
the A score of the input provided matrix
}
\description{
A_crit returns the A score for a specific experiment design. The A score is
the trace of whatever matrix is being optimized.
}
\examples{

M <- one_way_cov_B(error = 1, tau = 1, a = 10, n = 20)
score <- A_crit(M)

}
