% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{balanced_designs}
\alias{balanced_designs}
\title{Balanced Designs data}
\format{
\subsection{\code{who}}{

A data frame with 96 rows and 7 columns:
\describe{
\item{N}{total number of experiment runs}
\item{a}{number of groups in the random effect}
\item{n}{number of replicates per group}
\item{tau}{the ratio of the random effect variance component and the
residual error variance component}
\item{Criteria}{a character indicating the use of the A or D criteria for
the 'Score' columm}
\item{Score}{the score of the criteria for that row's experiment design}
\item{Cross.Score}{the score of the criteria not indicated in the
'Criteria' column for that row's experiment design}
...
}
}
}
\usage{
balanced_designs
}
\description{
A catalog of balanced designs using the one-way ANOVA model
}
\keyword{datasets}
