% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_B.R
\name{compare_designs_B}
\alias{compare_designs_B}
\title{compare_designs_B}
\usage{
compare_designs_B(designs)
}
\arguments{
\item{designs}{a data.frame object returned by the function
subset_designs_B()}
}
\value{
a ggplot object
}
\description{
compare_designs_B is used to create a plot of balanced designs all with the
same number of design points, but varying number of groups and values of tau.
}
\examples{

candidate_designs <- subset_designs_B(N = 100, criteria = "D")
compare_designs_B(designs = candidate_designs)

}
