% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_U.R
\name{compare_designs_U}
\alias{compare_designs_U}
\title{compare_designs_U}
\usage{
compare_designs_U(data, criteria)
}
\arguments{
\item{data}{a data.frame created using generate_designs_U()}

\item{criteria}{a character "D" or "A" to indicate the criteria designs are
scored on. Should match criteria used in the generate_designs_U() dataset}
}
\value{
a ggplot object
}
\description{
compare_designs_U creates a scatterplot of dataset index and relative
efficiency for a dataset of unbalanced designs created from
generate_designs_U(). A reference line is included to easily identify designs
that are at least 90\% efficient compared to the optimal design
}
\examples{
candidates <- generate_designs_U(N = 20, a = 5, sig_a_sq = 2, error = 1)
compare_designs_U(data = candidates, criteria = "D")
}
