% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_designs_B.R
\name{contour_designs_B}
\alias{contour_designs_B}
\title{contour_designs_B}
\usage{
contour_designs_B(
  ngroups = c(5, 10, 20, 30),
  nreps = c(5, 10, 20),
  taus = c(0.5, 1, 2, 5),
  error = 1,
  criteria = "D"
)
}
\arguments{
\item{ngroups}{a vector of integers indicating the number of groups to cross
with number of replications per group}

\item{nreps}{a vector of integers indicating the number of replications per
group to cross with number of groups}

\item{taus}{a vector of doubles indicating the number of values of tau to
cross with the number of crossed groups by replications}

\item{error}{a double indicating the estimated residual error of the model}

\item{criteria}{a character "D" or "A" indicating which design criteria to
use}
}
\value{
a list of two lists; the first being a list containing contour plots
that compare criteria scores and the second being a list containing contour
plots that compare relative efficiencies
}
\description{
contour_designs_B creates a list of size two in which both elements are a
list of contour plots. The first element is a list of contour plots comparing
criteria score across user specified ranges of designs. The second element is
a list of contour plots comparing relative efficiencies across user specified
ranges of designs. The relative efficiencies are calculated in comparison to
the design with the lowest criteria score for the range of designs on the
specific plot. The purpose of this tool is to allow users to compare how
running an experiment with a different number of groups or replications per
group can result in a comparable design performance. This can be useful if
running an experiment in a comparable way is more cost effective.
}
\details{
NOTE: This function will return a list of two lists of length
(ngroups)(nreps)(taus). It is best to save the output of this function to
a variable in your global environment. See example below.
}
\examples{

candidate_designs <- contour_designs_B(ngroups = c(10, 20),
                                        nreps = c(10, 20),
                                        taus = c(0.5, 1, 2),
                                        error = 1,
                                        criteria = "D")

}
