% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_crit.R
\name{D_crit}
\alias{D_crit}
\title{D_crit}
\usage{
D_crit(M)
}
\arguments{
\item{M}{a matrix object}
}
\value{
the D score of the input matrix
}
\description{
D_crit returns the D score for a specific experiment design. The D score is
the determinant of whatever matrix is being optimized.
}
\examples{

M <- general_variance_2VC(N = 20, n = 4, a = 5, sig_a_sq = 1, error_sq = 1)
score <- D_crit(M)

}
