% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_variance_2VC.R
\name{general_variance_2VC}
\alias{general_variance_2VC}
\title{general_variance_2VC}
\usage{
general_variance_2VC(N, n, a, sig_a_sq, error_sq)
}
\arguments{
\item{N}{an integer representing the total number of experimental runs}

\item{n}{either a single integer or a vector of integers representing the
number of runs per group. If a single integer is used, then the experiment is
treated as a balanced design. If a vector is used, the user can specify
unbalanced replications for the groups}

\item{a}{an integer representing the number of groups}

\item{sig_a_sq}{a double representing the estimate of \eqn{\sigma^2_A}}

\item{error_sq}{a double representing the estimate of \eqn{\sigma^2}}
}
\value{
the covariance matrix of the model's maximum likelihood estimators
}
\description{
general_variance_2VC returns the MLE covariance matrix for optimization
in a nested ANOVA model with two variance components: \eqn{\sigma^2_A} and
\eqn{\sigma^2}
}
\examples{
# Balanced design
general_variance_2VC(N = 8, n = 2, a = 4, sig_a_sq = 2, error_sq = 1)

# Unbalanced design
general_variance_2VC(N = 8, n = c(2, 3, 3), a = 3, sig_a_sq = 2, error_sq = 1)
}
