% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_designs_3VC_U.R
\name{generate_designs_3VC_U}
\alias{generate_designs_3VC_U}
\title{place holder}
\usage{
generate_designs_3VC_U(N, a, b, sig_a_sq, sig_b_sq, error_sq)
}
\arguments{
\item{N}{place holder}

\item{a}{place holder}

\item{b}{place holder}
}
\value{
place holder
}
\description{
place holder
}
