% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_designs_B.R
\name{generate_designs_B}
\alias{generate_designs_B}
\title{generate_designs_B}
\usage{
generate_designs_B(ngroups = c(5, 10), nreps = c(5, 10), taus = c(1, 2))
}
\arguments{
\item{ngroups}{a vector of integers specifying the number of groups}

\item{nreps}{a vector of integers specifying the number of replications per
group}

\item{taus}{a vector of doubles specifying the ratio of the variance components}
}
\value{
a dataframe of potential designs
}
\description{
generate_designs_B creates a dataframe of experiment designs with user
specified parameters.
}
\examples{

candidate_designs <- generate_designs_B(ngroups = c(3, 5), nreps = c(5, 10),
taus = c(1, 2))

}
