% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_two_way_designs.R
\name{generate_two_way_designs}
\alias{generate_two_way_designs}
\title{generate_two_way_designs}
\usage{
generate_two_way_designs(N, s, sig_a_sq, sig_b_sq, error_sq)
}
\arguments{
\item{N}{the total number of desired replicates in the experiment}

\item{s}{refers to desired class of experiments. An entry of 2 will produce
designs from class C_3_2 while an entry of 3 will produce designs from class
C_3_3}

\item{sig_a_sq}{a double representing the estimate of \eqn{\sigma^2_A}}

\item{sig_b_sq}{a double representing the estimate of \eqn{\sigma^2_B}}

\item{error_sq}{a double representing the estimate of \eqn{\sigma^2}}
}
\value{
a data.frame of all designs of size N from the selected class such
that all variance components can be estimated
}
\description{
generate_two_way_designs allows users to specify the parameters of a
two-way nested random effects design and the class of designs they would
like to select from to generate a data.frame of candidate designs. The
classes are named C_(3,2) and C_(3,3) in which the first subscript refers to
the number of variance components in the model and the second subscript
refers to the max number of sub-groups per level of alpha and max number of
replicates per sub-group.
}
\examples{
candidates <- generate_two_way_designs(N = 12, s = 2, sig_a_sq = 1,
                                       sig_b_sq = 1, error_sq = 1)

}
