% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way_dataset_doc.R
\docType{data}
\name{N12_0.1_0.1}
\alias{N12_0.1_0.1}
\alias{N12_0.1_0.5,}
\alias{N12_0.1_1,}
\alias{N12_0.1_2,}
\alias{N12_0.1_10,}
\alias{N12_0.5_0.1,}
\alias{N12_0.5_0.5,}
\alias{N12_0.5_1,}
\alias{N12_0.5_2,}
\alias{N12_0.5_10,}
\alias{N12_1_0.1,}
\alias{N12_1_0.5,}
\alias{N12_1_1,}
\alias{N12_1_2,}
\alias{N12_1_10,}
\alias{N12_2_0.1,}
\alias{N12_2_0.5,}
\alias{N12_2_1,}
\alias{N12_2_2,}
\alias{N12_2_10,}
\alias{N12_10_0.1,}
\alias{N12_10_0.5,}
\alias{N12_10_1,}
\alias{N12_10_2,}
\alias{N12_10_10,}
\alias{N24_0.1_0.1,}
\alias{N24_0.1_0.5,}
\alias{N24_0.1_1,}
\alias{N24_0.1_2,}
\alias{N24_0.1_10,}
\alias{N24_0.5_0.1,}
\alias{N24_0.5_0.5,}
\alias{N24_0.5_1,}
\alias{N24_0.5_2,}
\alias{N24_0.5_10,}
\alias{N24_1_0.1,}
\alias{N24_1_0.5,}
\alias{N24_1_1,}
\alias{N24_1_2,}
\alias{N24_1_10,}
\alias{N24_2_0.1,}
\alias{N24_2_0.5,}
\alias{N24_2_1,}
\alias{N24_2_2,}
\alias{N24_2_10,}
\alias{N24_10_0.1,}
\alias{N24_10_0.5,}
\alias{N24_10_1,}
\alias{N24_10_2,}
\alias{N24_10_10,}
\alias{N12_3_0.1_0.1,}
\alias{N12_3_0.1_0.5,}
\alias{N12_3_0.1_1,}
\alias{N12_3_0.1_2,}
\alias{N12_3_0.1_10,}
\alias{N12_3_0.5_0.1,}
\alias{N12_3_0.5_0.5,}
\alias{N12_3_0.5_1,}
\alias{N12_3_0.5_2,}
\alias{N12_3_0.5_10,}
\alias{N12_3_1_0.1,}
\alias{N12_3_1_0.5,}
\alias{N12_3_1_1,}
\alias{N12_3_1_2,}
\alias{N12_3_1_10,}
\alias{N12_3_2_0.1,}
\alias{N12_3_2_0.5,}
\alias{N12_3_2_1,}
\alias{N12_3_2_2,}
\alias{N12_3_2_10,}
\alias{N12_3_10_0.1,}
\alias{N12_3_10_0.5,}
\alias{N12_3_10_1,}
\alias{N12_3_10_2,}
\alias{N12_3_10_10}
\alias{N12_0.1_0.5}
\alias{N12_0.1_1}
\alias{N12_0.1_2}
\alias{N12_0.1_10}
\alias{N12_0.5_0.1}
\alias{N12_0.5_0.5}
\alias{N12_0.5_1}
\alias{N12_0.5_2}
\alias{N12_0.5_10}
\alias{N12_1_0.1}
\alias{N12_1_0.5}
\alias{N12_1_1}
\alias{N12_1_2}
\alias{N12_1_10}
\alias{N12_2_0.1}
\alias{N12_2_0.5}
\alias{N12_2_1}
\alias{N12_2_2}
\alias{N12_2_10}
\alias{N12_10_0.1}
\alias{N12_10_0.5}
\alias{N12_10_1}
\alias{N12_10_2}
\alias{N12_10_10}
\alias{N24_0.1_0.1}
\alias{N24_0.1_0.5}
\alias{N24_0.1_1}
\alias{N24_0.1_2}
\alias{N24_0.1_10}
\alias{N24_0.5_0.1}
\alias{N24_0.5_0.5}
\alias{N24_0.5_1}
\alias{N24_0.5_2}
\alias{N24_0.5_10}
\alias{N24_1_0.1}
\alias{N24_1_0.5}
\alias{N24_1_1}
\alias{N24_1_2}
\alias{N24_1_10}
\alias{N24_2_0.1}
\alias{N24_2_0.5}
\alias{N24_2_1}
\alias{N24_2_2}
\alias{N24_2_10}
\alias{N24_10_0.1}
\alias{N24_10_0.5}
\alias{N24_10_1}
\alias{N24_10_2}
\alias{N24_10_10}
\alias{N12_3_0.1_0.1}
\alias{N12_3_0.1_0.5}
\alias{N12_3_0.1_1}
\alias{N12_3_0.1_2}
\alias{N12_3_0.1_10}
\alias{N12_3_0.5_0.1}
\alias{N12_3_0.5_0.5}
\alias{N12_3_0.5_1}
\alias{N12_3_0.5_2}
\alias{N12_3_0.5_10}
\alias{N12_3_1_0.1}
\alias{N12_3_1_0.5}
\alias{N12_3_1_1}
\alias{N12_3_1_2}
\alias{N12_3_1_10}
\alias{N12_3_2_0.1}
\alias{N12_3_2_0.5}
\alias{N12_3_2_1}
\alias{N12_3_2_2}
\alias{N12_3_2_10}
\alias{N12_3_10_0.1}
\alias{N12_3_10_0.5}
\alias{N12_3_10_1}
\alias{N12_3_10_2}
\title{Datasets of two-way nested random effects designs generated from
\eqn{\mathcal{C}_{3,2}} and \eqn{\mathcal{C}_{3,3}}.}
\format{
\subsection{\verb{two-way datasets}}{

dataframes of designs from one of the two aforementioned classes with
prespecified number of observations and values for \eqn{\rho_1} and
\eqn{\rho_2}. The data files are titled such that the first identifier
(ex. N12) refers to the sample size, the second identifier (ex. 0.1) refers
to the value of \eqn{\rho_1}, and the third identifier (ex. 0.1) refers
to the value of \eqn{\rho_2}. Some files have four identifiers in which the
second identifier is "3". This means the dataset contains designs from
\eqn{\mathcal{C}_{3,3}} rather than \eqn{\mathcal{C}_{3,2}}. In such files,
the third and fourth identifier refer to the values of \eqn{\rho_1} and
\eqn{\rho_2}.

\describe{
\item{N}{Number of total observations}
\item{N_i_dot}{number of total observations within group i}
\item{N_ij}{number of observations within each subgroup}
\item{D_score}{The determinant of the inverse information matrix}
\item{Relative_D_Efficiency}{Efficiency compared to the D optimal design}
\item{A_score}{The trace of the inverse information matrix}
\item{Relative_A_Efficiency}{Efficiency compared to the A optimal design}
}
}
}
\usage{
N12_0.1_0.1

N12_0.1_0.5

N12_0.1_1

N12_0.1_2

N12_0.1_10

N12_0.5_0.1

N12_0.5_0.5

N12_0.5_1

N12_0.5_2

N12_0.5_10

N12_1_0.1

N12_1_0.5

N12_1_1

N12_1_2

N12_1_10

N12_2_0.1

N12_2_0.5

N12_2_1

N12_2_2

N12_2_10

N12_10_0.1

N12_10_0.5

N12_10_1

N12_10_2

N12_10_10

N24_0.1_0.1

N24_0.1_0.5

N24_0.1_1

N24_0.1_2

N24_0.1_10

N24_0.5_0.1

N24_0.5_0.5

N24_0.5_1

N24_0.5_2

N24_0.5_10

N24_1_0.1

N24_1_0.5

N24_1_1

N24_1_2

N24_1_10

N24_2_0.1

N24_2_0.5

N24_2_1

N24_2_2

N24_2_10

N24_10_0.1

N24_10_0.5

N24_10_1

N24_10_2

N24_10_10

N12_3_0.1_0.1

N12_3_0.1_0.5

N12_3_0.1_1

N12_3_0.1_2

N12_3_0.1_10

N12_3_0.5_0.1

N12_3_0.5_0.5

N12_3_0.5_1

N12_3_0.5_2

N12_3_0.5_10

N12_3_1_0.1

N12_3_1_0.5

N12_3_1_1

N12_3_1_2

N12_3_1_10

N12_3_2_0.1

N12_3_2_0.5

N12_3_2_1

N12_3_2_2

N12_3_2_10

N12_3_10_0.1

N12_3_10_0.5

N12_3_10_1

N12_3_10_2
}
\description{
Datasets of two-way nested random effects designs generated from
\eqn{\mathcal{C}_{3,2}} and \eqn{\mathcal{C}_{3,3}}.
}
\keyword{datasets}
