% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way_cov_B.R
\name{one_way_cov_B}
\alias{one_way_cov_B}
\title{one_way_cov_B}
\usage{
one_way_cov_B(error = 1, tau = 1, a, n)
}
\arguments{
\item{error}{the residual error variance component}

\item{tau}{a double representing the ratio of the random effect variance
component and the residual error variance component \deqn{\frac{\sigma^2_\alpha}{\sigma^2}}}

\item{a}{an integer representing the number of groups in a one-way ANOVA
experiment design}

\item{n}{an integer number of replications per group in a one-way ANOVA
experiment design}
}
\value{
the maximum likelihood estimator covariance matrix for a user
specified balanced one-way ANOVA model.
}
\description{
one_way_cov_B provides the asymptotic covariance matrix for
the maximum likelihood estimators of a one-way ANOVA experiment design with a
single random effect. This matrix represents the inverse of the information
matrix of the user's model and should be minimized in optimal design of
experiments.
}
\examples{

M <- one_way_cov_B(error = 1, tau = 1, a = 5, n = 5)

}
