% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way_cov_U.R
\name{one_way_cov_U}
\alias{one_way_cov_U}
\title{one_way_cov_U}
\usage{
one_way_cov_U(N, a, n_i, vc, error)
}
\arguments{
\item{N}{an integer number of total design points in the experiment}

\item{a}{an integer number of groups in the experiment}

\item{n_i}{a vector of doubles representing the number of replications per group}

\item{vc}{a double representing the estimate of the variance component}

\item{error}{a double representing the residual mean error}
}
\value{
the maximum likelihood estimator covariance matrix for a user
specified unbalanced one-way ANOVA model.
}
\description{
one_way_cov_U provides the asymptotic covariance matrix for the maximum
likelihood estimators of an unbalanced one-way ANOVA experiment design with a
single random effect. This matrix represents the inverse of the information
matrix of the user's model and should be minimized in optimal design of
experiments.
}
\examples{

u_info <- u_infomat(N = 10, a = 4, n_i = c(1,2,3,4), vc = 2, error = 1)

}
