% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{plot_design}
\usage{
plot_design(n, g, sig_a, error, criteria)
}
\arguments{
\item{n}{a vector of integers representing the number of replications per group}

\item{g}{an integer representing the number of groups}

\item{sig_a}{a double indicating the value of \deqn{\sigma^2_A}}

\item{error}{a double indicating the value of \deqn{\sigma^2}}

\item{criteria}{a character "D" or "A" indicating the criteria to score on}
}
\value{
a tree diagram showing the structure of the experiment
}
\description{
plot_design
}
\examples{
plot_design(n = c(3,4,5), g = 3, sig_a = 2, error = 1, criteria = "D")
}
