% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_designs_B.R
\name{subset_designs_B}
\alias{subset_designs_B}
\title{subset_designs_B}
\usage{
subset_designs_B(data, N, criteria)
}
\arguments{
\item{data}{a dataframe of designs generated using generate_designs_B()}

\item{N}{an integer value representing the total number of experiment design
points}

\item{criteria}{a character value representing the desired design criteria}
}
\value{
a data.frame object of balanced designs meeting user specifications
}
\description{
subset_designs_B returns a dataset of balanced designs that have a user
specified number of design points. This dataset allows the user to compare
various designs of the same size across different numbers of groups,
replications, and ratio of variance components (tau)
}
\examples{

designs <- subset_designs_B(N = 100, criteria = "D")
}
