% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_2VC.R
\name{V_2VC}
\alias{V_2VC}
\title{V_2VC}
\usage{
V_2VC(n, a, sig_a_sq, error_sq)
}
\arguments{
\item{n}{either a single integer or a vector of integers representing the
number of runs per group. If a single integer is used, then the experiment is
treated as a balanced design. If a vector is used, the user can specify
unbalanced replications for the groups}

\item{a}{an integer representing the number of groups}

\item{sig_a_sq}{a double representing the estimate of \deqn{\sigma^2_A}}

\item{error_sq}{a double representing the estimate of \deqn{\sigma^2}}
}
\value{
the variance matrix of a 2-way nested anova model
}
\description{
V_2VC returns the variance matrix for a one-way nested ANOVA model that has
a two variance components: \deqn{\sigma^2_A} and \deqn{\sigma^2}.
}
\examples{
# Balanced design
V_2VC(n = 4, a = 2, sig_a_sq = 2, error_sq = 1)

# Unbalanced design
V_2VC(n = c(3, 5), a = 2, sig_a_sq = 2, error_sq = 1)
}
