% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_3VC.R
\name{V_3VC}
\alias{V_3VC}
\title{V_3VC}
\usage{
V_3VC(N, n_a, n_b, sig_a_sq, sig_b_sq, error_sq)
}
\arguments{
\item{N}{an integer representing the total number of experimental runs}

\item{n_a}{integers representing the total number of reps in each group at
the first level of the experiment}

\item{n_b}{integers representing the total number of reps in each group at
the second level of the experiment}

\item{sig_a_sq}{a double representing the estimate of \deqn{\sigma^2_A}}

\item{sig_b_sq}{a double representing the estimate of \deqn{\sigma^2_B}}

\item{error_sq}{a double representing the estimate of \deqn{\sigma^2}}
}
\value{
the variance matrix of a 2-way nested anova model
}
\description{
V_3VC returns the variance matrix for a 2-way nested ANOVA model that has
a total of three variance components: \deqn{\sigma^2_A}, \deqn{\sigma^2_B},
and \deqn{\sigma^2}
}
\examples{
# Balanced design
V_3VC(N = 8, n_a = c(4, 4), n_b = c(2, 2, 2, 2), sig_a_sq = 2, sig_b_sq = 3, error_sq = 1)

# Unbalanced design
V_3VC(N = 9, n_a = c(4, 5), n_b = c(2, 3, 1, 2, 1), sig_a_sq = 2, sig_b_sq = 3, error_sq = 1)
}
