% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{verify_2_way_info}
\alias{verify_2_way_info}
\title{verify_2_way_info is a helper function that calculates the covariance matrix
of maximum likelihood estimators of a balanced two-way nested random effects model.
See "Searle, S. R., George Casella, and Charles E. McCulloch. Variance
Components. Wiley Series in Probability and Statistics. Hoboken, NJ:
Wiley, 2006." page 158.}
\usage{
verify_2_way_info(n, a, b, sig_a_sq, sig_b_sq, error_sq)
}
\arguments{
\item{n}{integer number of replications per sub-group}

\item{a}{integer number of groups}

\item{b}{integer number of sub-groups per group}

\item{sig_a_sq}{estimate of \deqn{\sigma^2_A}}

\item{sig_b_sq}{estimate of \deqn{\sigma^2_B}}

\item{error_sq}{estimate of \deqn{\sigma^2}}
}
\value{
a matrix object that is the covariance matrix of the maximum
likelihood estimators for the variance components of the model.
}
\description{
verify_2_way_info is a helper function that calculates the covariance matrix
of maximum likelihood estimators of a balanced two-way nested random effects model.
See "Searle, S. R., George Casella, and Charles E. McCulloch. Variance
Components. Wiley Series in Probability and Statistics. Hoboken, NJ:
Wiley, 2006." page 158.
}
\examples{
#Balanced Design
verify_2_way_info(n = 2, a = 2, b = 2, sig_a_sq = 1, sig_b_sq = 1, error_sq = 1)

}
