% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_crit.R
\name{A_crit}
\alias{A_crit}
\title{A_crit}
\usage{
A_crit(M)
}
\arguments{
\item{M}{a matrix object}
}
\value{
the A score of the input matrix
}
\description{
A_crit returns the A score for a specific experiment design. The A score is
the trace of whatever matrix is being optimized.
}
\examples{

M <- general_variance_2VC(N = 20, n = 4, a = 5, sig_a_sq = 1, error_sq = 1)
score <- A_crit(M)

}
