% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_B.R
\name{compare_designs_B}
\alias{compare_designs_B}
\title{compare_designs_B}
\usage{
compare_designs_B(designs, criteria)
}
\arguments{
\item{designs}{a data.frame object of designs of size N returned by the
function subset_designs_B()}
}
\value{
a dashboard of ggplot objects
}
\description{
compare_designs_B is used to create a dashboard of comparison plots for
designs of the same size N. The top left plot compares the score of the
designs across values of tau. The top right plot compares the relative
efficiency of all designs of size N despite the value of tau. The remaining
plots compare the relative efficiency of designs for specific values of tau.
}
\examples{

candidate_designs <- subset_designs_B(N = 100, criteria = "D")
compare_designs_B(designs = candidate_designs)

}
