% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_crit.R
\name{D_crit}
\alias{D_crit}
\title{D_crit}
\usage{
D_crit(M)
}
\arguments{
\item{M}{a matrix object}
}
\value{
the D score of the input provided matrix
}
\description{
D_crit returns the D score for a specific experiment design. The D score is
the determinant of whatever matrix is being optimized.
}
\examples{

M <- one_way_balanced_MLE_covariance(error = 1, tau = 1, a = 10, n = 20)
score <- D_crit(M)

}
